// src/lib/sdk/models/CreateBOMDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateBOMDto = {
    /**
     * ID of the product this BOM belongs to
     */
    productId: string;
    /**
     * Name of the Bill of Materials
     */
    name: string;
    /**
     * Description of the Bill of Materials
     */
    description?: string;
};