// src/lib/sdk/models/UpdateBOMDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateBOMDto = {
    /**
     * Name of the Bill of Materials
     */
    name?: string;
    /**
     * Description of the Bill of Materials
     */
    description?: string;
    /**
     * Whether this BOM is active
     */
    isActive?: boolean;
};