// src/lib/sdk/models/CreateBOMItemDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateBOMItemDto = {
    /**
     * ID of the stock item to include in this BOM
     */
    stockItemId: string;
    /**
     * Quantity of this stock item required in the BOM
     */
    quantity: number;
};